added static casting to stuff

+++ AlgorithmsFFLib/src/TargetDecoyCandidatePairManager.cpp
@@ -223,21 +223,21 @@ namespace {
                 if (ionLableInfo.second.contains("^2")) {
 
                     if (ionLableInfo.first >= firstIndexToMutate) {
-                        ms2IonDecoy.mz += nTermDeltaMassCharge2;
+                        ms2IonDecoy.mz += static_cast<float>(nTermDeltaMassCharge2);
                     }
 
                     if (ionLableInfo.first >= secondIndexToMutate) {
-                        ms2IonDecoy.mz += cTermDeltaMassCharge2;
+                        ms2IonDecoy.mz += static_cast<float>(cTermDeltaMassCharge2);
                     }
                 }
                 else {
 
                     if (ionLableInfo.first >= firstIndexToMutate) {
-                        ms2IonDecoy.mz += nTermDeltaMass;
+                        ms2IonDecoy.mz += static_cast<float>(nTermDeltaMass);
                     }
 
                     if (ionLableInfo.first >= secondIndexToMutate) {
-                        ms2IonDecoy.mz += cTermDeltaMass;
+                        ms2IonDecoy.mz += static_cast<float>(cTermDeltaMass);
                     }
                 }
             }
@@ -247,21 +247,21 @@ namespace {
                 if (ionLableInfo.second.contains("^2")) {
 
                     if (ionLableInfo.first >= firstIndexToMutate) {
-                        ms2IonDecoy.mz += cTermDeltaMassCharge2;
+                        ms2IonDecoy.mz += static_cast<float>(cTermDeltaMassCharge2);
                     }
 
                     if (ionLableInfo.first >= secondIndexToMutate) {
-                        ms2IonDecoy.mz += nTermDeltaMassCharge2;
+                        ms2IonDecoy.mz += static_cast<float>(nTermDeltaMassCharge2);
                     }
                 }
                 else {
 
                     if (ionLableInfo.first >= firstIndexToMutate) {
-                        ms2IonDecoy.mz += cTermDeltaMass;
+                        ms2IonDecoy.mz += static_cast<float>(cTermDeltaMass);
                     }
 
                     if (ionLableInfo.first >= secondIndexToMutate) {
-                        ms2IonDecoy.mz += nTermDeltaMass;
+                        ms2IonDecoy.mz += static_cast<float>(nTermDeltaMass);
                     }
                 }
             }
